refactor : cqrs적용에 따른 test 코드 수정
