--- uppick-auction-service/src/test/java/org/oneog/uppick/auction/domain/product/entity/ProductTest.java
@@ -1,9 +1,0 @@
-package org.oneog.uppick.auction.domain.product.entity;
-
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-@ExtendWith(MockitoExtension.class)
-public class ProductTest {
-
-}
