+++ uppick-auction-service/src/test/java/org/oneog/uppick/auction/domain/product/query/controller/ProductQueryControllerRestDocsTest.java
@@ -1,4 +1,4 @@
-package org.oneog.uppick.auction.domain.product.controller;
+package org.oneog.uppick.auction.domain.product.query.controller;
 
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.BDDMockito.*;
@@ -15,9 +15,8 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 import org.junit.jupiter.api.Test;
-import org.oneog.uppick.auction.domain.product.command.model.dto.request.ProductRegisterRequest;
-import org.oneog.uppick.auction.domain.product.command.model.dto.request.SearchProductRequest;
 import org.oneog.uppick.auction.domain.product.common.enums.ProductSearchSortType;
+import org.oneog.uppick.auction.domain.product.query.model.dto.request.SearchProductRequest;
 import org.oneog.uppick.auction.domain.product.query.model.dto.response.ProductBiddingInfoResponse;
 import org.oneog.uppick.auction.domain.product.query.model.dto.response.ProductDetailResponse;
 import org.oneog.uppick.auction.domain.product.query.model.dto.response.ProductSellingInfoResponse;
@@ -25,7 +24,7 @@ import org.oneog.uppick.auction.domain.product.query.model.dto.response.ProductS
 import org.oneog.uppick.auction.domain.product.query.model.dto.response.PurchasedProductInfoResponse;
 import org.oneog.uppick.auction.domain.product.query.model.dto.response.SearchProductInfoResponse;
 import org.oneog.uppick.auction.domain.product.query.model.dto.response.SoldProductInfoResponse;
-import org.oneog.uppick.auction.domain.product.service.ProductInternalService;
+import org.oneog.uppick.auction.domain.product.query.service.ProductQueryService;
 import org.oneog.uppick.auction.support.auth.WithMockAuthMember;
 import org.oneog.uppick.auction.support.restdocs.RestDocsBase;
 import org.oneog.uppick.common.dto.AuthMember;
@@ -36,70 +35,19 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.MediaType;
-import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.restdocs.payload.JsonFieldType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
-import org.springframework.web.multipart.MultipartFile;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-@WebMvcTest(ProductController.class)
-class ProductControllerRestDocsTest extends RestDocsBase {
+@WebMvcTest(ProductQueryController.class)
+class ProductQueryControllerRestDocsTest extends RestDocsBase {
 
 	@MockitoBean
-	private ProductInternalService productInternalService;
+	private ProductQueryService productQueryService;
 
 	@Autowired
 	private ObjectMapper objectMapper;
-
-	@Test
-	@WithMockAuthMember(memberId = 10L, memberNickname = "tester")
-	void registerProduct_정상적인입력_상품등록성공() throws Exception {
-
-		ProductRegisterRequest request = ProductRegisterRequest.builder()
-			.name("테스트 상품")
-			.description("상품 설명")
-			.categoryId(1L)
-			.startBid(1000L)
-			.endAt(LocalDateTime.now().plusDays(7))
-			.build();
-
-		MockMultipartFile productPart = new MockMultipartFile(
-			"product", "product", "application/json",
-			objectMapper.writeValueAsBytes(request));
-
-		MockMultipartFile imagePart = new MockMultipartFile(
-			"image", "image.jpg", "image/jpeg",
-			"dummy-image-content".getBytes());
-
-		willDoNothing().given(productInternalService)
-			.registerProduct(any(ProductRegisterRequest.class), any(MultipartFile.class), eq(10L));
-
-		this.mockMvc.perform(multipart("/api/v1/products")
-				.file(productPart)
-				.file(imagePart)
-				.header("Authorization", "Bearer token")
-				.contentType(MediaType.MULTIPART_FORM_DATA)
-				.accept(MediaType.APPLICATION_JSON))
-			.andExpect(status().isOk())
-			.andDo(document("product-register-product",
-				requestHeaders(
-					headerWithName("Authorization").description("JWT 액세스 토큰 (Bearer {token})")),
-				requestParts(
-					partWithName("product").description("상품 등록 요청 JSON"),
-					partWithName("image").description("상품 대표 이미지 파일")),
-				requestPartFields("product",
-					fieldWithPath("name").type(JsonFieldType.STRING).description("상품 이름"),
-					fieldWithPath("description").type(JsonFieldType.STRING).description("상품 설명"),
-					fieldWithPath("categoryId").type(JsonFieldType.NUMBER).description("카테고리 ID"),
-					fieldWithPath("startBid").type(JsonFieldType.NUMBER).description("시작 입찰가"),
-					fieldWithPath("endAt").type(JsonFieldType.STRING).description("마감 일시 (ISO-8601)")),
-				responseFields(
-					fieldWithPath("success").type(JsonFieldType.BOOLEAN).description("요청 성공 여부"),
-					fieldWithPath("message").type(JsonFieldType.STRING).description("응답 메시지"),
-					fieldWithPath("data").type(JsonFieldType.NULL).description("응답 데이터 (여기서는 null)"))));
-
-	}
 
 	@Test
 	@WithMockAuthMember(memberId = 10L, memberNickname = "tester")
@@ -121,7 +69,7 @@ class ProductControllerRestDocsTest extends RestDocsBase {
 			.sellerName("sellerName")
 			.build();
 
-		given(productInternalService.getProductInfoById(eq(productId), any(AuthMember.class)))
+		given(productQueryService.getProductInfoById(eq(productId), any(AuthMember.class)))
 			.willReturn(response);
 
 		this.mockMvc.perform(get("/api/v1/products/{productId}", productId)
@@ -163,7 +111,7 @@ class ProductControllerRestDocsTest extends RestDocsBase {
 			.currentBidPrice(1_500L)
 			.build();
 
-		given(productInternalService.getProductSimpleInfoById(eq(productId)))
+		given(productQueryService.getProductSimpleInfoById(eq(productId)))
 			.willReturn(response);
 
 		this.mockMvc.perform(get("/api/v1/products/{productId}/simple-info", productId)
@@ -207,7 +155,7 @@ class ProductControllerRestDocsTest extends RestDocsBase {
 
 		Page<SoldProductInfoResponse> page = new PageImpl<>(contents, Pageable.ofSize(20), 2);
 
-		given(productInternalService.getSoldProductInfosByMemberId(eq(10L), any(Pageable.class)))
+		given(productQueryService.getSoldProductInfosByMemberId(eq(10L), any(Pageable.class)))
 			.willReturn(page);
 
 		this.mockMvc.perform(get("/api/v1/products/sold/me")
@@ -253,7 +201,7 @@ class ProductControllerRestDocsTest extends RestDocsBase {
 
 		Page<PurchasedProductInfoResponse> page = new PageImpl<>(contents, Pageable.ofSize(20), 2);
 
-		given(productInternalService.getPurchasedProductInfoByMemberId(eq(10L), any(Pageable.class)))
+		given(productQueryService.getPurchasedProductInfoByMemberId(eq(10L), any(Pageable.class)))
 			.willReturn(page);
 
 		this.mockMvc.perform(get("/api/v1/products/purchased/me")
@@ -299,7 +247,7 @@ class ProductControllerRestDocsTest extends RestDocsBase {
 				.build());
 		Page<ProductBiddingInfoResponse> page = new PageImpl<>(contents, Pageable.ofSize(10), 2);
 
-		given(productInternalService.getBiddingProductInfoByMemberId(eq(10L), any(Pageable.class)))
+		given(productQueryService.getBiddingProductInfoByMemberId(eq(10L), any(Pageable.class)))
 			.willReturn(page);
 
 		this.mockMvc.perform(get("/api/v1/products/bidding/me")
@@ -346,14 +294,14 @@ class ProductControllerRestDocsTest extends RestDocsBase {
 				.build());
 		Page<ProductSellingInfoResponse> page = new PageImpl<>(contents, Pageable.ofSize(10), 2);
 
-		given(productInternalService.getSellingProductInfoByMemberId(eq(10L), any(Pageable.class)))
+		given(productQueryService.getSellingProductInfoByMemberId(eq(10L), any(Pageable.class)))
 			.willReturn(page);
 
 		this.mockMvc.perform(get("/api/v1/products/selling/me")
 				.header("Authorization", "Bearer token")
 				.accept(MediaType.APPLICATION_JSON))
 			.andExpect(status().isOk())
-			.andDo(document("product-get-bidding-products",
+			.andDo(document("product-get-selling-products",
 				requestHeaders(
 					headerWithName("Authorization").description("JWT 액세스 토큰 (Bearer {token})")),
 				responseFields(
@@ -397,7 +345,7 @@ class ProductControllerRestDocsTest extends RestDocsBase {
 
 		Page<SearchProductInfoResponse> page = new PageImpl<>(List.of(item), PageRequest.of(0, 10), 1);
 
-		given(productInternalService.searchProduct(any(SearchProductRequest.class))).willReturn(page);
+		given(productQueryService.searchProduct(any(SearchProductRequest.class))).willReturn(page);
 
 		// when & then
 		mockMvc.perform(
